fix(brain-core): move PackTunable to brain-core, eliminate pack→pack dep
pack-memory imported PackTunable/ParameterSpace/ParameterDef from
pack-brain, creating a pack→pack dependency that violates ADR-017.

Move the tunable contract types to khive-brain-core (adding khive-runtime
dep for PackRuntime supertrait + RuntimeError). pack-brain's tunable.rs
becomes a thin re-export shim. pack-memory now imports from brain-core
directly — no more khive-pack-brain in its Cargo.toml.
