+++ crates/khive-pack-memory/Cargo.toml
@@ -17,7 +17,6 @@ khive-fusion = { version = "0.2.8", path = "../khive-fusion" }
 khive-retrieval = { version = "0.2.8", path = "../khive-retrieval" }
 khive-score = { version = "0.2.8", path = "../khive-score" }
 khive-brain-core = { version = "0.2.8", path = "../khive-brain-core" }
-khive-pack-brain = { version = "0.2.8", path = "../khive-pack-brain" }
 khive-vamana = { version = "0.2.8", path = "../khive-vamana" }
 inventory = { workspace = true }
 khive-storage = { version = "0.2.8", path = "../khive-storage" }
