+++ crates/khive-pack-memory/src/tunable.rs
@@ -1,7 +1,7 @@
 //! Brain-tunable parameter surface for the memory pack's recall scoring pipeline.
 
 use khive_brain_core::BalancedRecallState;
-use khive_pack_brain::tunable::{PackTunable, ParameterDef, ParameterSpace};
+use khive_brain_core::{PackTunable, ParameterDef, ParameterSpace};
 use khive_runtime::RuntimeError;
 use serde_json::Value;
